+++ packages/app/src/cli/prompts/init/init.ts
@@ -52,16 +52,6 @@ export const templates = {
       options: {
         javascript: {branch: 'javascript', label: 'JavaScript'},
         typescript: {branch: 'main', label: 'TypeScript'},
-        ...(process.env.POLARIS_UNIFIED && {
-          javascriptPolarisEarlyAccess: {
-            branch: 'polaris-docs-2025-js',
-            label: 'JavaScript (Polaris Release Candidate)',
-          },
-          typescriptPolarisEarlyAccess: {
-            branch: 'polaris-docs-2025',
-            label: 'TypeScript (Polaris Release Candidate)',
-          },
-        }),
       },
     },
   } as Template,