+++ packages/store/vite.config.ts
@@ -1,3 +1,0 @@
-import config from '../../configurations/vite.config'
-
-export default config(__dirname)