+++ packages/plugin-did-you-mean/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shopify/plugin-did-you-mean
 
+## 3.85.0
+
+### Patch Changes
+
+- Updated dependencies [b855ec1]
+- Updated dependencies [a883d5b]
+- Updated dependencies [74fc99c]
+- Updated dependencies [be916a1]
+- Updated dependencies [1416fe1]
+- Updated dependencies [6497461]
+  - @shopify/cli-kit@3.85.0
+
 ## 3.84.0
 
 ### Patch Changes